Removed Terminal Option
Removed the option for the user to enter code through the terminal. The goal was to primarily use that function for testing but it is not necessary and testing is working fine with .txt files. Also removed the switch case that was used to read the user's choice between file entry and terminal entry.