Number Parsing
Created a class called Compiler to hold all the functions involved in parsing and outputting code.
Added ability for compiler to identify numbers and whether they are integers or floats.
Changed input3.txt to feature an integer for testing purposes. input2.txt contains a float. The program is able to identify these two types of numbers successfully.