Pause After Error
Added a system("PAUSE"); after an error message for when a file is not opened properly. This is because when using the .exe application, inputting an invalid file path closed the application right away. This would leave the user confused and unsure how to fix the issue. Adding a pause lets the user read the error message and know what to do.